refactor: change the responses to custom responses
