+++ internal/adapters/api/handler/permission.go
@@ -1,9 +1,11 @@
-package api
+package handler
 
 import (
 	"github.com/gin-gonic/gin"
+	"github.com/iBoBoTi/project_boiler_plate/internal/adapters/api/response"
 	"github.com/iBoBoTi/project_boiler_plate/internal/core/domain"
 	"github.com/iBoBoTi/project_boiler_plate/internal/core/ports"
+	"net/http"
 )
 
 type permissionHandler struct {
@@ -19,40 +21,40 @@ func NewPermissionHandler(permissionService ports.PermissionService) ports.Permi
 func (h *permissionHandler) CreatePermission(c *gin.Context) {
 	permission := domain.Permission{}
 	if err := c.ShouldBindJSON(&permission); err != nil {
-		c.JSON(400, gin.H{"error": err.Error()})
+		response.JSON(c, "invalid_request_body", http.StatusBadRequest, nil, []string{err.Error()})
 		return
 	}
 	if err := h.permissionService.CreatePermission(&permission); err != nil {
-		c.JSON(400, gin.H{"error": err.Error()})
+		response.JSON(c, "invalid_input", http.StatusBadRequest, nil, []string{err.Error()})
 		return
 	}
-	c.JSON(200, gin.H{"message": "permission successfully created"})
+	response.JSON(c, "permission successfully created", http.StatusCreated, nil, nil)
 }
 
 func (h *permissionHandler) GetPermissionByID(c *gin.Context) {
 	permissionID := c.Param("id")
 	permission, err := h.permissionService.GetPermissionByID(permissionID)
 	if err != nil {
-		c.JSON(400, gin.H{"error": err.Error()})
+		response.JSON(c, "invalid_input", http.StatusBadRequest, nil, []string{err.Error()})
 		return
 	}
-	c.JSON(200, gin.H{"permission": permission})
+	response.JSON(c, "permission gotten", http.StatusOK, permission, nil)
 }
 
 func (h *permissionHandler) GetAllPermissions(c *gin.Context) {
 	permissions, err := h.permissionService.GetAllPermissions()
 	if err != nil {
-		c.JSON(400, gin.H{"error": err.Error()})
+		response.JSON(c, "invalid_input", http.StatusBadRequest, nil, []string{err.Error()})
 		return
 	}
-	c.JSON(200, gin.H{"permissions": permissions})
+	response.JSON(c, "permissions gotten", http.StatusOK, permissions, nil)
 }
 
 func (h *permissionHandler) DeletePermission(c *gin.Context) {
 	id := c.Param("id")
 	if err := h.permissionService.DeletePermission(id); err != nil {
-		c.JSON(400, gin.H{"error": err.Error()})
+		response.JSON(c, "invalid_input", http.StatusBadRequest, nil, []string{err.Error()})
 		return
 	}
-	c.JSON(200, gin.H{"message": "permission successfully deleted"})
+	response.JSON(c, "permission successfully deleted", http.StatusOK, nil, nil)
 }
